+++ server/app.js
@@ -1,5 +1,7 @@
+import createError from 'http-errors';
 import express from 'express';
 import path from 'path';
+import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
 import indexRouter from './routes/index.js';
 import authRouter from './routes/auth.js';
@@ -16,6 +18,7 @@ app.set('view engine', 'ejs');
 app.use(morgan('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
+app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
@@ -23,7 +26,7 @@ app.use('/auth', authRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
-  res.sendStatus(404);
+  next(createError(404));
 });
 
 // error handler
