+++ server/bin/www
@@ -4,10 +4,10 @@
  * Module dependencies.
  */
 
-var app = require('../app');
-var debug = require('debug')('server:server');
-var http = require('http');
-
+ import app from '../app.js';
+ import http from 'http';
+ import Debug from 'debug';
+ var debug = Debug("server:server");
 /**
  * Get port from environment and store in Express.
  */
