+++ server/package.json
@@ -4,7 +4,8 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "start": "nodemon app"
+    "start": "nodemon ./bin/www",
+    "test": "nodemon app"
   },
   "dependencies": {
     "bcrypt": "^5.0.1",
@@ -13,6 +14,7 @@
     "dotenv": "^10.0.0",
     "ejs": "~2.6.1",
     "express": "~4.16.1",
+    "http": "0.0.1-security",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
     "mysql2": "^2.3.0",
