+++ server/routes/auth.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import * as authController from '../controller/auth.js';
+import * as authService from '../services/auth.js';
 
 const router = express.Router();
 router.get('/signup', function(req, res, next) {
   res.render('signup', { title: 'Todaily' });
 });
 
-router.post('/signup', authController.signup);
+router.post('/signup', authService.signup);
 
 export default router;
